requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.4.0
billiard==4.2.4
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==42.0.8
dj-rest-auth==6.0.0
Django==5.1.15
django-allauth==0.63.6
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-timezone-field==7.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
et_xmlfile==2.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.4.2
openpyxl==3.1.5
phonenumbers==8.13.55
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2024.2
PyYAML==6.0.3
redis==5.0.8
referencing==0.37.0
reportlab==4.0.9
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
